택배 송장 제조기 [STEP1] gogochang 	- class -> struct 수정 	- adress -> address 오탈자 수정 	- private 변경 	- 파라미터 개행스타일로 변경 	- 하위 구조체 접근 메서드 방식 추가
